refactor: Typos, unnecessary types, file location
