+++ app/locales/en/messages.po
@@ -1106,7 +1106,7 @@ msgstr "Understood, I'll proceed cautiously."
 
 #: app/configurator/components/add-dataset-dialog.tsx
 msgid "dataset.search.caution.body"
-msgstr "The linking of different datasets carries risks such as data inconsistencies, scalability issues, and unexpected correlations. Be sure to use to merge datasets only when you are confident that data should be merge together."
+msgstr "The linking of different datasets carries risks such as data inconsistencies, scalability issues, and unexpected correlations. Be sure to use to merge datasets only when you are confident that data should be merged together."
 
 #: app/browser/dataset-browse.tsx
 #: app/configurator/components/add-dataset-dialog.tsx
